Replace value-prefixed format with CBOR Seq
CBOR is self-delimiting, and RFC 8742 describes the ability to construct
CBOR Sequences, which are just concatenated top-level CBOR structures.

Since fixed-size CBOR structures are length-prefixed, there is no need
for us to further length-prefix them. We can just concat.

The latest version of serde_ipld_dagcbor supports the ability to decode
CBOR Sequences with new library features. We're using that to sidestep
the need for a custom format. This makes SZDT archives valid CBOR Seq.